requirements.txt
Django==1.7
elasticsearch==1.6.0
psycopg2==2.6
djangorestframework==3.1.3
requests==2.7.0
django-nose==1.4.1
coverage==3.7.1
